requirements.txt
# Core FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database & ORM
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Data Collection & Processing
yfinance==0.2.36
pandas==2.2.0
numpy==1.26.3
requests==2.31.0
aiohttp==3.9.1
asyncio==3.4.3

# Caching & Performance
redis==5.0.1
python-multipart==0.0.6

# Visualization
plotly==5.18.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# AI Integration (Free Models)
openai==1.10.0
anthropic==0.8.1
google-generativeai==0.3.2

# Background Tasks & Scheduling
celery==5.3.6
redis==5.0.1
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
pytz==2023.3.post1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
